+++ CHANGELOG.md
@@ -37,6 +37,8 @@
 
 ### Fixed
 
+- [#6467](https://github.com/ChainSafe/forest/pull/6467) Fixed: `Filecoin.EthGetBlockByNumber` now only supports retrieving a block by its block number or a special tag.
+
 ## Forest v0.31.1 "Quadrantids"
 
 This is a non-mandatory release for all node operators. It includes the support for more V2 API's and a few critical API fixes.